Fixing the random float generation for Longitude and Latitude initial db seeding

+++ PickUpApi/Models/Helpers/Location.cs
@@ -7,7 +7,7 @@ namespace PickUpApi.Models.Helpers
         //TODO: Need to find the appropriate data format for this.
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int LocationId { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
     }
 }
